make styles folder and move css files in it
