+++ frontend/src/Components/Header/Header.tsx
@@ -1,16 +1,12 @@
-import React from 'react';
-import './Header.css'; 
+import React from "react";
+import "../../Styles/Header.css";
 
+const Header = () => {
+  return (
+    <div className="header">
+      <button className="help-button">Help</button>
+    </div>
+  );
+};
 
-const Header = ()=> {
-    return(
-        <div className='header'>
-            <button className='help-button'>Help</button>
-        </div>
-    )
-}
-
-export default Header; 
-
-
-
+export default Header;
