+++ frontend/src/Components/NavBar/NavBar.tsx
@@ -1,32 +1,39 @@
-import React from 'react';
-import './NavBar.css'; 
+import React from "react";
+import "../../Styles/NavBar.css";
 import { BsCalendarCheck, BsCardList, BsFillPersonFill } from "react-icons/bs";
 
+const NavBar = () => {
+  const onClickToday = async () => {
+    window.location.href = "/today";
+  };
 
-const NavBar = ()=> {
+  const onClickDeliveries = async () => {
+    window.location.href = "/tasks";
+  };
 
-    const onClickToday = async () => {
-        window.location.href = '/today';
-    };
+  const onClickMe = async () => {
+    window.location.href = "/profile";
+  };
 
-    const onClickDeliveries = async () => {
-        window.location.href = '/tasks';
-    };
+  return (
+    <div className="navBar">
+      <hr></hr>
+      <div className="horizontalBar">
+        <button onClick={onClickToday}>
+          {" "}
+          <BsCalendarCheck className="icon" /> <br /> Today
+        </button>
+        <button onClick={onClickDeliveries}>
+          {" "}
+          <BsCardList className="icon" /> <br /> My Deliveries
+        </button>
+        <button onClick={onClickMe}>
+          {" "}
+          <BsFillPersonFill className="icon" /> <br /> Me
+        </button>
+      </div>
+    </div>
+  );
+};
 
-    const onClickMe = async () => {
-        window.location.href = '/profile';
-    };
-
-    return(
-        <div className='navBar'>
-            <hr ></hr>
-            <div className='horizontalBar'>
-                <button onClick={onClickToday}> <BsCalendarCheck className='icon'/> <br/> Today</button>
-                <button onClick={onClickDeliveries}> <BsCardList className='icon'/> <br/> My Deliveries</button>
-                <button onClick={onClickMe}> <BsFillPersonFill className='icon'/> <br/> Me</button>
-            </div>
-        </div>
-    )
-}
-
-export default NavBar; 
+export default NavBar;
