+++ frontend/src/Containers/TodayContainer.tsx
@@ -3,15 +3,13 @@ import {
   DeliveryTimeline,
   NoDeliveries,
 } from "../Components/Today/DeliveryTimeline";
-import "../Components/Today/Delivery.css";
+import "../Styles/Delivery.css";
 
 const TodayContainer = () => {
   let noDeliveries = false;
   return (
     <div>
-      <Box
-        className="center"
-      >
+      <Box className="center">
         <Box
           className="welcome"
           sx={{ display: "flex", flexDirection: "column", "margin-top": "15%" }}
@@ -36,12 +34,11 @@ const TodayContainer = () => {
           >
             Today's Deliveries
           </Typography>
-          {!noDeliveries && <DeliveryTimeline />}{" "} {/* display the screen of the timeline */}
+          {!noDeliveries && <DeliveryTimeline />}{" "}
+          {/* display the screen of the timeline */}
         </Box>
       </Box>
-
       {noDeliveries && <NoDeliveries />}{" "}
-
       {/* display the screen for no deliveries  */}
       <Box display="flex" width={"100%"} justifyContent="center">
         <Button sx={{ backgroundColor: "#33BE41" }} variant="contained">
